requirements.txt
pysqlite3-binary==0.5.2
streamlit==1.49.1
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.10
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
sentence-transformers==5.1.0
chromadb==1.0.20
jira==3.10.5
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
python-dotenv==1.1.1
pypdf==6.0.0
docx2txt==0.9
pillow==11.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic[email]